# runtime/packages/client/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Scrapyomama Runtime API
  version: 0.1.0
  license:
    name: MIT
    identifier: MIT
servers:
  - url: /
# Par défaut : session d'interface (cookie) ou clé d'API (Bearer, scopes de 13 § 8). Les routes publiques posent `security: []` ;
# les routes réservées à la session (registre `auth: 'session'`) posent `security: [{sessionCookie: []}]` : jamais une clé d'API
# sur une route d'administration (13 § 8). Le test de dérive compare ce mode d'authentification au registre du serveur.
security:
  - sessionCookie: []
  - apiKey: []
...
paths:
...
  /api/apis/{slug}/schedules:
    get:
      operationId: listSchedules
      tags: [schedules]
      summary: Planifications d'une API (08 § 5)
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: Planifications.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleList'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    post:
      operationId: createSchedule
      tags: [schedules]
      summary: Crée une planification (fréquence minimale 1 minute)
      parameters:
        - $ref: '#/components/parameters/Slug'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ScheduleWrite'
      responses:
        '201':
          description: Planification créée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
  /api/apis/{slug}/schedules/{id}:
    get:
      operationId: getSchedule
      tags: [schedules]
      summary: Une planification, avec ses prochaines exécutions calculées côté serveur
      parameters:
        - $ref: '#/components/parameters/Slug'
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Planification.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '401':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    patch:
      operationId: updateSchedule
      tags: [schedules]
      summary: Modifie une planification (dont la suspension)
      parameters:
        - $ref: '#/components/parameters/Slug'
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SchedulePatch'
      responses:
        '200':
          description: Planification à jour.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
    delete:
      operationId: deleteSchedule
      tags: [schedules]
      summary: Supprime une planification
      parameters:
        - $ref: '#/components/parameters/Slug'
        - $ref: '#/components/parameters/Id'
      responses:
        '204':
          description: Supprimée.
        '401':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-sy.session
      description: Session d'interface (13 § 5) ; `sy.session` sans le préfixe `__Host-` quand PUBLIC_URL est en http (développement).
    apiKey:
      type: http
      scheme: bearer
      description: Clé d'API `sy_live_…` à scopes (13 § 8) ; jamais de scope d'administration.
    deviceToken:
      type: http
      scheme: bearer
      description: Jeton d'appareil de l'extension (07 § 1), lié à (utilisateur, appareil) ; jamais accepté hors de /api/extension.
    metricsToken:
      type: http
      scheme: bearer
      description: METRICS_TOKEN (14 § 3), distinct des clés d'API.
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    Slug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9-]{0,62}$'
...
  responses:
    Error:
      description: "Erreur au format commun `{ error: { code, message } }` (le code est stable, jamais localisé)."
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    ApiStatus:
      type: string
      enum: [enquete, sain, warning, reparation, erreur, action_requise, bloquee]
...
    ScheduleRules:
      type: object
      additionalProperties: false
      properties:
        only_if_tunnel_online:
          type: boolean
        window:
          type: [object, 'null']
          additionalProperties: false
          required: [start, end]
          properties:
            start:
              type: string
              pattern: '^\d{2}:\d{2}$'
            end:
              type: string
              pattern: '^\d{2}:\d{2}$'
            days:
              type: array
              items:
                type: integer
                minimum: 0
                maximum: 6
        max_runs_per_day:
          type: [integer, 'null']
          minimum: 1
        skip_if_status_in:
          description: Défaut `[erreur, action_requise, bloquee]` (on ne sollicite pas un site qui a refusé).
          type: array
          items:
            $ref: '#/components/schemas/ApiStatus'
        dedup_key:
          type: [string, 'null']
        diff:
          type: [string, 'null']
          enum: [new, changed, removed, all, null]
        alert_on:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [new_items, status_change, error]
    ScheduleWrite:
      type: object
      additionalProperties: false
      required: [cron, timezone, input]
      properties:
        cron:
          type: string
          maxLength: 100
        timezone:
          type: string
          maxLength: 64
        input:
          description: Entrée du run ; variables datées `{{today}}`, `{{yesterday}}`.
          type: object
          additionalProperties: true
        overlap:
          type: string
          enum: [skip, queue, allow]
          default: skip
        missed:
          type: string
          enum: [once, skip]
          default: once
        rules:
          $ref: '#/components/schemas/ScheduleRules'
        enabled:
          type: boolean
    SchedulePatch:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        cron:
          type: string
          maxLength: 100
        timezone:
          type: string
          maxLength: 64
        input:
          type: object
          additionalProperties: true
        overlap:
          type: string
          enum: [skip, queue, allow]
        missed:
          type: string
          enum: [once, skip]
        rules:
          $ref: '#/components/schemas/ScheduleRules'
        enabled:
          type: boolean
    Schedule:
      type: object
      additionalProperties: false
      required: [id, api_slug, cron, timezone, input, overlap, missed, rules, enabled, next_runs, created_at]
      properties:
        id:
          type: string
          format: uuid
        api_slug:
          type: string
        cron:
          type: string
        timezone:
          type: string
        input:
          type: object
          additionalProperties: true
        overlap:
          type: string
          enum: [skip, queue, allow]
        missed:
          type: string
          enum: [once, skip]
        rules:
          $ref: '#/components/schemas/ScheduleRules'
        enabled:
          type: boolean
        paused_reason:
          type: [string, 'null']
        next_runs:
          description: Prochaines exécutions calculées côté serveur (06 § 1, aucun composant cron dans la console).
          type: array
          items:
            type: string
            format: date-time
        created_at:
          type: string
          format: date-time
    ScheduleList:
      type: object
      additionalProperties: false
      required: [schedules]
      properties:
        schedules:
          type: array
          items:
            $ref: '#/components/schemas/Schedule'
    # ——— Webhooks (Standard Webhooks, 08 § 5) ———
...